requirements.txt
pytest-playwright
playwright
pytest-xdist
pytest-html
pytest-check